requirements.txt
asgiref==3.2.10
Django==3.0.7
pytz==2020.1
sqlparse==0.3.1
django-ckeditor==5.9.0
Pillow==7.1.2
requests==2.24.0
requests-oauthlib==1.3.0
urllib3==1.25.9
psycopg2-binary==2.8.5
gunicorn==20.0.4
django-environ==0.4.5
djangorestframework==3.11.0
python-nginx==1.5.3
cloudflare-ddns==1.3.0
django-bootstrap4==2.1.1
social-auth-app-django==3.1.0
pyTelegramBotAPI==3.7.2
pytube3==9.6.4
python-telegram-bot==12.8
celery==4.4.7
redis==3.5.3
django-celery-beat==2.0.0
